Polish the rest of the details for footer and move the footer to containers/footer
