+++ app/root.tsx
@@ -16,7 +16,7 @@ import { honeypot } from './utils/honeypot.server';
 import { HoneypotProvider } from 'remix-utils/honeypot/react';
 import type { HoneypotInputProps } from 'remix-utils/honeypot/server';
 import { ThemeSwitch } from '~/components/theme-switch';
-import { Footer } from '~/components/footer';
+import { Footer } from '~/containers/footer';
 import { SunOutlineIcon } from './icons/outline/sun';
 import { MoonOutlineIcon } from './icons/outline/moon';
 import { themeSessionResolver } from './utils/theme.server';
@@ -88,7 +88,7 @@ export function App() {
 
                 <div className="h-screen overflow-y-scroll">
                     <Outlet />
-                    <Footer className='mt-[250px] text-center pb-[5rem]' />
+                    <Footer className='mt-[120px] text-center mb-[5rem] py-[2rem]' />
                 </div>
                 <ScrollRestoration />
                 <Scripts />
